+++ wsdiscovery/message.py
@@ -1,31 +1,30 @@
-"""
-Top-level SOAP message creation and parsing
-"""
+"""Functions to serialize and deserialize messages between SOAP envelope & string representations"""
 
 import io
-from .namespaces import NS_A, NS_S
+from .namespaces import NS_ADDRESSING, NS_SOAPENV
 from .actions import *
 from xml.dom import minidom
 
 
 def createSOAPMessage(env):
-    "construct a raw SOAP XML string, given a prepared SoapEnvelope object"
-    if env.getAction() == ACTION_PROBE:
+    "serialize SOAP envelopes into XML strings"
+
+    if env.getAction() == NS_ACTION_PROBE:
         return createProbeMessage(env)
-    if env.getAction() == ACTION_PROBE_MATCH:
+    if env.getAction() == NS_ACTION_PROBE_MATCH:
         return createProbeMatchMessage(env)
-    if env.getAction() == ACTION_RESOLVE:
+    if env.getAction() == NS_ACTION_RESOLVE:
         return createResolveMessage(env)
-    if env.getAction() == ACTION_RESOLVE_MATCH:
+    if env.getAction() == NS_ACTION_RESOLVE_MATCH:
         return createResolveMatchMessage(env)
-    if env.getAction() == ACTION_HELLO:
+    if env.getAction() == NS_ACTION_HELLO:
         return createHelloMessage(env)
-    if env.getAction() == ACTION_BYE:
+    if env.getAction() == NS_ACTION_BYE:
         return createByeMessage(env)
 
 
 def parseSOAPMessage(data, ipAddr):
-    "parse raw XML data string, return a (minidom) xml document"
+    "deserialize XML message strings into SOAP envelope objects"
 
     try:
         dom = minidom.parseString(data)
@@ -33,20 +32,20 @@ def parseSOAPMessage(data, ipAddr):
         #print('Failed to parse message from %s\n"%s": %s' % (ipAddr, data, ex), file=sys.stderr)
         return None
 
-    if dom.getElementsByTagNameNS(NS_S, "Fault"):
+    if dom.getElementsByTagNameNS(NS_SOAPENV, "Fault"):
         #print('Fault received from %s:' % (ipAddr, data), file=sys.stderr)
         return None
 
-    soapAction = dom.getElementsByTagNameNS(NS_A, "Action")[0].firstChild.data.strip()
-    if soapAction == ACTION_PROBE:
+    soapAction = dom.getElementsByTagNameNS(NS_ADDRESSING, "Action")[0].firstChild.data.strip()
+    if soapAction == NS_ACTION_PROBE:
         return parseProbeMessage(dom)
-    elif soapAction == ACTION_PROBE_MATCH:
+    elif soapAction == NS_ACTION_PROBE_MATCH:
         return parseProbeMatchMessage(dom)
-    elif soapAction == ACTION_RESOLVE:
+    elif soapAction == NS_ACTION_RESOLVE:
         return parseResolveMessage(dom)
-    elif soapAction == ACTION_RESOLVE_MATCH:
+    elif soapAction == NS_ACTION_RESOLVE_MATCH:
         return parseResolveMatchMessage(dom)
-    elif soapAction == ACTION_BYE:
+    elif soapAction == NS_ACTION_BYE:
         return parseByeMessage(dom)
-    elif soapAction == ACTION_HELLO:
+    elif soapAction == NS_ACTION_HELLO:
         return parseHelloMessage(dom)