+++ wsdiscovery/namespaces.py
@@ -1,10 +1,34 @@
-"""
-Base namespaces
-"""
+"""SOAP & WS-Discovery XML namespaces used by the package"""
 
-NS_A = "http://schemas.xmlsoap.org/ws/2004/08/addressing"
-NS_D = "http://schemas.xmlsoap.org/ws/2005/04/discovery"
-NS_S = "http://www.w3.org/2003/05/soap-envelope"
 
+#: Addressing namespace
+NS_ADDRESSING = "http://schemas.xmlsoap.org/ws/2004/08/addressing"
+
+#: Discovery namespace
+NS_DISCOVERY = "http://schemas.xmlsoap.org/ws/2005/04/discovery"
+
+#: SOAP envelope namespace
+NS_SOAPENV = "http://www.w3.org/2003/05/soap-envelope"
+
+NS_ADDRESS_ALL = "urn:schemas-xmlsoap-org:ws:2005:04:discovery"
+NS_ADDRESS_UNKNOWN = "http://schemas.xmlsoap.org/ws/2004/08/addressing/role/anonymous"
+
+#: :term:`Bye` message namespace
+NS_ACTION_BYE = "http://schemas.xmlsoap.org/ws/2005/04/discovery/Bye"
+
+#: :term:`Hello` message namespace
+NS_ACTION_HELLO = "http://schemas.xmlsoap.org/ws/2005/04/discovery/Hello"
+
+#: :term:`Probe` message namespace
+NS_ACTION_PROBE = "http://schemas.xmlsoap.org/ws/2005/04/discovery/Probe"
+
+#: :term:`Probe match` message namespace
+NS_ACTION_PROBE_MATCH = "http://schemas.xmlsoap.org/ws/2005/04/discovery/ProbeMatches"
+
+#: :term:`Resolve` message namespace
+NS_ACTION_RESOLVE = "http://schemas.xmlsoap.org/ws/2005/04/discovery/Resolve"
+
+#: :term:`Resolve match` message namespace
+NS_ACTION_RESOLVE_MATCH = "http://schemas.xmlsoap.org/ws/2005/04/discovery/ResolveMatches"